// src/app/business-logic/model/events/taskMetric.ts
/**
 * events
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.11
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface TaskMetric {
    /**
     * Task ID
     */
    task: string;
    /**
     * Metric name
     */
    metric: string;
}